+++ test/command-subst.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+echo "I'm in $(pwd)"
+echo "I'm in `pwd`"