+++ test/cond-exec.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+echo "now" && echo "then"
+true || echo "false"
+false || echo "true"