+++ test/quotes.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+NAME="John"
+echo "Hi $NAME"  #=> Hi John
+echo 'Hi $NAME'  #=> Hi $NAME