+++ test/vars.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+NAME="John"
+echo $NAME
+echo "$NAME"
+echo "${NAME}!"